Split responses up into book, instrument, etc.
