+++ parquet/include/pairs_sink.hpp
@@ -1,7 +1,9 @@
 #pragma once
 
 #include "io.hpp"
-#include "responses.hpp"
+
+#include <header.hpp>
+#include <pair.hpp>
 
 #include <arrow/api.h>
