+++ test/parse_instrument_snapshot.cpp
@@ -1,7 +1,7 @@
-#include "responses.hpp"
+#include "instrument.hpp"
 
-#include <iostream>
 #include <simdjson.h>
+#include <iostream>
 
 int main(int argc, char* argv[]) {
   if (argc != 2) {
