+++ test/unit/level_book_test.cpp
@@ -1,7 +1,6 @@
 #include <doctest/doctest.h>
 
 #include <level_book.hpp>
-#include <responses.hpp>
 
 #include <simdjson.h>
